Rename scrapy/squeue.py to scrapy/squeues.py
