+++ scrapy/squeues.py
